% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dennys.R
\docType{data}
\name{dennys}
\alias{dennys}
\title{Locations of Denny's restaurants}
\format{
A tibble with 1643 rows and 3 variables:
\describe{
\item{address}{}
\item{city}{}
\item{state}{}
\item{zip}{}
\item{longitude}{}
\item{latitude}{}
}
}
\source{
Need source.
}
\usage{
dennys
}
\description{
Need description
}
\examples{
dennys
}
\keyword{datasets}
