% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheries.R
\docType{data}
\name{fisheries}
\alias{fisheries}
\title{Fishing industry by country}
\format{A tibble with 75 rows and 3 variables:
\describe{
\item{country}{Name of country}
\item{capture}{Tonnage from capture, including fish, crustaceans, molluscs, etc.}
\item{aquaculture}{Tonnage from aquaculture}
}}
\source{
https://en.wikipedia.org/wiki/Fishing_industry_by_country
}
\usage{
fisheries
}
\description{
World fisheries harvest for 2005. The tonnage from capture and aquaculture
is listed by country. Countries which harvested less than 100,000 tons are
not included.
}
\examples{
fisheries
}
\keyword{datasets}
