% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss.R
\docType{data}
\name{gss}
\alias{gss}
\title{Deprecated}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2867 rows and 9 columns.
}
\usage{
gss
}
\description{
This dataset is now deprecated. Use \code{\link{gss16}} instead.
}
\seealso{
\code{\link{gss16}}
}
\keyword{datasets}
