% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laquinta.R
\docType{data}
\name{laquinta}
\alias{laquinta}
\title{Locations of La Quinta Inn motels}
\format{
A tibble with 909 rows and 3 variables:
\describe{
\item{address}{}
\item{city}{}
\item{state}{}
\item{zip}{}
\item{longitude}{}
\item{latitude}{}
}
}
\source{
Need source.
}
\usage{
laquinta
}
\description{
Need description
}
\examples{
laquinta
}
\keyword{datasets}
