% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lego_sales.R
\docType{data}
\name{lego_sales}
\alias{lego_sales}
\title{Sales of Lego sets (simulated)}
\format{
A tibble with 620 rows and 14 variables:
\describe{
\item{first_name}{First name of customer}
\item{last_name}{Last name of customer}
\item{age}{Age of customer}
\item{phone_number}{Phone number of customer}
\item{set_id}{Set ID of lego set purchased}
\item{number}{Item number of lego set purchased}
\item{theme}{Theme of lego set purchased}
\item{subtheme}{Sub theme of lego set purchased}
\item{year}{Year of purchase}
\item{name}{ Name of lego set purchased}
\item{pieces}{Number of pieces of legos in set purchased}
\item{us_price}{Price of set purchase in US Dollars}
\item{image_url}{Image URL of lego set purchased}
\item{quantity}{Quantity of lego set(s) purchased}
}
}
\usage{
lego_sales
}
\description{
Simulated data based on real prices for Lego sales in 2018 in the United States.
Names and hobbies have been randomly generated
}
\examples{

library(dplyr)

lego_sales \%>\%
  count(first_name, sort = TRUE) \%>\%
  top_n(3)

}
\keyword{datasets}
