% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbreweries.R
\docType{data}
\name{ncbreweries}
\alias{ncbreweries}
\title{Breweries in North Carolina}
\format{
A tibble with 251 rows and 7 variables:
\describe{
\item{name}{Name of brewery}
\item{city}{City where brewery is located}
\item{type}{Type of brewery (Microbrewery, Client Brewer, Brewpub,
Brewpub/Brewery, Commercial Brewery)}
\item{beercount}{Number of beers brewed at the brewery}
\item{est}{Year brewery opened}
\item{status}{Status of brewery (Active or Closed)}
\item{url}{URL of brewery page on RateBeer.com}
}
}
\source{
https://www.ratebeer.com/breweries/north\%20carolina/33/213/
}
\usage{
ncbreweries
}
\description{
This dataset is scraped from RateBeer.com on April 18, 2018, and it includes
a list of breweries in North Carolina as of that date.
}
\examples{
instructors
}
\keyword{datasets}
