% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucbadmit.R
\docType{data}
\name{ucbadmit}
\alias{ucbadmit}
\title{UC Berkeley admissions}
\format{
A tibble with 4526 rows and 3 variables:
\describe{
\item{admit}{Admission status of the applicant: Admitted or Rejected}
\item{gender}{Gender of applicant: Male or Female}
\item{dept}{Anonymized department: A, B, C, D, E, or F}
}
}
\source{
R Core Team (2018). R: A language and environment for
statistical computing. R Foundation for Statistical Computing,
Vienna, Austria. URL https://www.R-project.org/.
}
\usage{
ucbadmit
}
\description{
This dataset is a flat (two-dimensional) version of the UCBAdmissions
dataset from the datasets package.
}
\examples{
ucbadmit
}
\keyword{datasets}
