% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss.R
\docType{data}
\name{gss}
\alias{gss}
\title{GSS - General Social Survey}
\format{
A tibble with 2,867 rows and 9 variables:
\describe{
\item{harass5}{"Over the past five years, have you been harassed by your superiors or co-workers at your job, for example, have you experienced any bullying, physical or psychological abuse?"}
\item{emailmin}{Number of minutes spent on email weekly, extra to the hours in emailhrs (e.g. emailmin = 30 for 2.5 hours on email).}
\item{emailhr}{Number of hours spent on email weekly.}
\item{educ}{Number of years in education.}
\item{polviews}{Political views. Possible answers are Extremely liberal, Liberal, Slightly liberal, Moderate, Slghtly conservative, Conservative, Extrmly conservative.}
\item{advfront}{"Even if it brings no immediate benefits, scientific research that advances the frontiers of knowledge is necessary and should be supported by the federal government."}
\item{snapchat}{Whether respondant uses Snapchat or not.}
\item{instagram}{Whether respondant uses Instagram or not.}
\item{wrkstat}{Work status.}
}
}
\source{
https://gss.norc.org/Get-The-Data
}
\usage{
gss
}
\description{
The GSS gathers data on contemporary American society in order to monitor and
explain trends and constants in attitudes, behaviors, and attributes.
Hundreds of trends have been tracked since 1972. In addition, since the GSS
adopted questions from earlier surveys, trends can be followed for up to 70 years.
The GSS contains a standard core of demographic, behavioral, and attitudinal
questions, plus topics of special interest. Among the topics covered are civil
liberties, crime and violence, intergroup tolerance, morality, national spending
priorities, psychological well-being, social mobility, and stress and traumatic events.
}
\details{
The data come from the \href{https://gss.norc.org/Get-The-Data}{NORC at the University of Chicago}

The data has been modified to better serve the goals of introductory data science education.
Only 9 of the original 935 variables have been selected for this dataset.
}
\examples{

library(ggplot2)

ggplot(gss, aes(x = emailhr)) +
  geom_histogram(binwidth = 2)

}
\keyword{datasets}
